+++ src/itemAction/index.tsx
@@ -1,5 +1,9 @@
-import React from 'react'
-import Item from '_utils/item'
+import React, { PureComponent } from 'react'
+import cc from 'classcat'
+
+import Loader from 'loader'
+import Item, { ItemStatus } from '_utils/item'
+import prefix from '_utils'
 
 export interface ItemActionProps {
   readonly tag?: JSX.Element
@@ -7,38 +11,79 @@ export interface ItemActionProps {
   readonly href?: string | JSX.Element
   readonly action?: string
   readonly subLabel?: string
+  readonly status?: ItemStatus
   readonly leftAddon?: React.ReactNode
   readonly onClick?: (event: React.MouseEvent<HTMLElement>) => void
   readonly onBlur?: (event: React.FocusEventHandler<HTMLElement>) => void
   readonly onFocus?: (event: React.FocusEventHandler<HTMLElement>) => void
   readonly onMouseDown?: (event: React.MouseEvent<HTMLElement>) => void
+  readonly onDoneAnimationEnd?: () => void
 }
 
-const ItemAction = ({
-  className,
-  action,
-  subLabel,
-  leftAddon,
-  tag,
-  href,
-  onClick,
-  onBlur,
-  onFocus,
-  onMouseDown,
-}: ItemActionProps) => (
-  <Item
-    highlighted
-    className={className}
-    leftAddon={leftAddon}
-    leftTitle={action}
-    leftBody={subLabel}
-    href={href}
-    onClick={onClick}
-    onBlur={onBlur}
-    onFocus={onFocus}
-    onMouseDown={onMouseDown}
-    tag={tag || <button />}
-  />
-)
+class ItemAction extends React.PureComponent<ItemActionProps> {
+  static defaultProps: Partial<ItemActionProps> = {
+    status: ItemStatus.DEFAULT,
+  }
+
+  static STATUS = ItemStatus
+
+  render() {
+    const {
+      className,
+      action,
+      subLabel,
+      leftAddon,
+      tag,
+      href,
+      status,
+      onClick,
+      onBlur,
+      onFocus,
+      onMouseDown,
+      onDoneAnimationEnd,
+    } = this.props
+
+    let rightAddon
+
+    if (status === ItemStatus.LOADING) {
+      rightAddon = (
+        <Loader
+          className={cc(prefix({ chevron: true }))}
+          size={24}
+          onDoneAnimationEnd={onDoneAnimationEnd}
+          inline
+        />
+      )
+    }
+    if (status === ItemStatus.CHECKED) {
+      rightAddon = (
+        <Loader
+          className={cc(prefix({ chevron: true }))}
+          size={24}
+          onDoneAnimationEnd={onDoneAnimationEnd}
+          done
+          inline
+        />
+      )
+    }
+
+    return (
+      <Item
+        highlighted
+        className={className}
+        leftAddon={leftAddon}
+        leftTitle={action}
+        leftBody={subLabel}
+        rightAddon={rightAddon}
+        href={href}
+        onClick={onClick}
+        onBlur={onBlur}
+        onFocus={onFocus}
+        onMouseDown={onMouseDown}
+        tag={tag || <button />}
+      />
+    )
+  }
+}
 
 export default ItemAction