+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[NEW]** Added `status` prop to `ItemAction`
 - **[FIX]** Fixed button with `CHECKED` status in `ButtonGroup`
 - **[DEV]** Added typescript check in pre-commit hook
 - [...]